moved media_data
