+++ setup.py
@@ -5,7 +5,7 @@ except ImportError:
 
 setup(
     name="gradio",
-    version="2.9.0b",
+    version="2.9.0b1",
     include_package_data=True,
     description="Python library for easily interacting with trained machine learning models",
     author="Abubakar Abid, Ali Abid, Ali Abdalla, Dawood Khan, Ahsen Khaliq, Pete Allen, Ömer Faruk Özdemir",
