+++ test/test_inputs.py
@@ -10,7 +10,7 @@ import pandas
 import PIL
 
 import gradio as gr
-from gradio.test_data import media_data
+from gradio import media_data
 
 os.environ["GRADIO_ANALYTICS_ENABLED"] = "False"
