+++ test/test_processing_utils.py
@@ -8,7 +8,7 @@ import numpy as np
 from PIL import Image
 
 import gradio as gr
-from gradio.test_data import media_data
+from gradio import media_data
 
 os.environ["GRADIO_ANALYTICS_ENABLED"] = "False"
